Code: Pi Calculator Program Version: 0.1.0.0
/*
Name: Douglas Oak

Student# :not listed for security as this is a public repository

Algorithm Description: This program uses a for loop to calculate ever increasingly 
accurate estimations of Pi. This can be done by changing the value of the "num_calculations"
variable. This variable controls how many fractions the initial starting value of 4 is 
multiplied against

No functions or external packages were used as part of this program

Version: 0.1.0.0

*/