Code: Lab 2 Wind Chill Temperature Calculator
Name: Douglas Oak

Student# :Not available

Algorithm Description: The purpose of this program is to temperature, wind speed and relative humidity 
data from the user and use those factors to calculate wind chill temperature. The program uses a while
loop to keep the program running and check for imporper entries. 

No functions or external packages were used as part of this program

Version: 0.1.0.0